refactor(ComponentTable): simplify prop typings
